+++ src/Product/product.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class productService {
+  getProduct() : string[]{
+    return ['article1','article2']
+  }
+}
